refactor: split `actor` module
